passes through 'persist' option
git-svn-id: http://192.168.1.10/repos/toto@3053 b53939d4-f08c-4181-96d4-b201aa9c61c4

+++ Toto/trunk/DatabaseServer/src/com/azquo/spreadsheet/DSSpreadsheetService.java
@@ -185,7 +185,7 @@ public class DSSpreadsheetService {
     }
 
     // create a file to import from a populated region in the spreadsheet
-    private static int importDataFromSpreadsheet(AzquoMemoryDBConnection azquoMemoryDBConnection, CellsAndHeadingsForDisplay cellsAndHeadingsForDisplay, String user) throws Exception {
+    private static int importDataFromSpreadsheet(AzquoMemoryDBConnection azquoMemoryDBConnection, CellsAndHeadingsForDisplay cellsAndHeadingsForDisplay, String user, boolean persist) throws Exception {
         //write the column headings and data to a temporary file, then import it
         String fileName = "temp_" + user;
         File temp = File.createTempFile(fileName + ".csv", "csv");
@@ -245,6 +245,9 @@ public class DSSpreadsheetService {
         UploadedFile uploadedFile = new UploadedFile(tempPath, Collections.singletonList("csv-" + cellsAndHeadingsForDisplay.getRegion()), false);
         DSImportService.readPreparedFile(azquoMemoryDBConnection, uploadedFile);
         // persist no longer automatic on importing so just do it here
+        if (!persist) {
+            return uploadedFile.getNoValuesAdjusted().get();
+        }
         new Thread(azquoMemoryDBConnection::persist).start();
         if (!temp.delete()) {// see no harm in this here. Delete on exit has a problem with Tomcat being killed from the command line. Why is intelliJ shirty about this?
             System.out.println("Unable to delete " + temp.getPath());
@@ -295,7 +298,7 @@ public class DSSpreadsheetService {
             azquoMemoryDBConnection.setProvenance(user, StringLiterals.IN_SPREADSHEET, reportName, context);
             if ((cellsAndHeadingsForDisplay.getRowHeadings().size()== 0 || cellsAndHeadingsForDisplay.getColumnHeadings().size() == 0) && cellsAndHeadingsForDisplay.getData().size() > 0) {
                 // todo - cen we get the number of values modified???
-                numberOfValuesModified = importDataFromSpreadsheet(azquoMemoryDBConnection, cellsAndHeadingsForDisplay, user);
+                numberOfValuesModified = importDataFromSpreadsheet(azquoMemoryDBConnection, cellsAndHeadingsForDisplay, user, persist);
                 if (persist) {
                     azquoMemoryDBConnection.persist();
                 }
